Move IOException to shared
